Fix bug: Add failure signal handle

+++ include/dev_lib/log/log_harness.h
@@ -12,7 +12,7 @@ public:
   ~LogHarness();
 
 private:
-  static void handleSingal(const char* data, size_t size);
+  static void handleSingal(const char* data, int size);
 };
 
 #endif
